+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 	- Alternating colors for rows.
 	- Visibility of anchor sprite is now taken into account.
 - zrcp: Improved stability of socket connection.
+- Disabled "cspect" support.
 
 # 1.2.5
 - Improved zsim performance when stepping over macro.